Port all test cases under transaction layer
We decide to get rid of Scala completely.

Port the codes.

+++ scalechain-transaction/src/main/kotlin/io/scalechain/blockchain/transaction/CoinAmount.kt
@@ -4,8 +4,8 @@ package io.scalechain.blockchain.transaction
   *
   * @param value The amount of coin value.
   */
-data class CoinAmount(val value : java.math.BigInteger) {
-  constructor( value : Long) : this(java.math.BigInteger.valueOf(value))
+data class CoinAmount(val value : java.math.BigDecimal) {
+  constructor( value : Long) : this(java.math.BigDecimal.valueOf(value))
   /** Return the coin amount in coin units.
     *
     * @return The coin units calculated from the CoinAmount. In Bitcoin, the units of coin is satoshi.
@@ -18,7 +18,7 @@ data class CoinAmount(val value : java.math.BigInteger) {
   companion object {
     /** How many units does a coin have?
      */
-    val ONE_COIN_IN_UNITS = java.math.BigInteger.valueOf(100000000L)
+    val ONE_COIN_IN_UNITS = java.math.BigDecimal.valueOf(100000000L)
 
     /** Get CoinAmount from the units of coin. In Bitcoin, the units of coin is satoshi.
      *
@@ -26,7 +26,7 @@ data class CoinAmount(val value : java.math.BigInteger) {
      * @return
      */
     fun from(coinUnits : Long) : CoinAmount {
-      return CoinAmount( java.math.BigInteger.valueOf(coinUnits) / ONE_COIN_IN_UNITS )
+      return CoinAmount( java.math.BigDecimal.valueOf(coinUnits) / ONE_COIN_IN_UNITS )
     }
   }
 }
